refactor check changed pages
