+++ src/utils/accessibilityTest.ts
@@ -3,7 +3,6 @@ export const accessibilityTest = async (
   config?: Record<string, unknown>
 ): Promise<void> => {
   if (typeof window !== 'undefined' && process.env.NODE_ENV !== 'production') {
-    console.log('Accessibility tests will run.');
     const axe = await import('@axe-core/react');
     const ReactDOM = await import('react-dom');
     axe.default(App, ReactDOM, 1000, config);
